+++ SearsCatalog/Config/PluginConfig.cs
@@ -13,8 +13,15 @@ public static class PluginConfig {
   public static ConfigEntry<int> BuildHudPanelColumns { get; private set; }
   public static ConfigEntry<Vector2> BuildHudPanelPosition { get; private set; }
 
+  public static ConfigEntry<float> BuildHudPanelScrollSensitivity { get; private set; }
+
   public static void BindConfig(ConfigFile config) {
-    IsModEnabled = config.BindInOrder("_Global", "isModEnabled", true, "Globally enable or disable this mod.");
+    IsModEnabled =
+        config.BindInOrder(
+            "_Global",
+            "isModEnabled",
+            true,
+            "Globally enable or disable this mod.");
 
     BuildHudPanelRows =
         config.BindInOrder(
@@ -44,5 +51,15 @@ public static class PluginConfig {
             "BuildHud.Panel position relative to center of the screen.");
 
     BuildHudPanelPosition.OnSettingChanged(BuildHudController.SetBuildPanelPosition);
+
+    BuildHudPanelScrollSensitivity =
+        config.BindInOrder(
+            "BuildHud.Panel",
+            "buildHudPanelScrollSensitivity",
+            1400f,
+            "BuildHud.Panel scroll-sensitivity when using the mouse-wheel.",
+            new AcceptableValueRange<float>(0f, 3000f));
+
+    BuildHudPanelScrollSensitivity.OnSettingChanged(BuildHudController.SetBuildPanelScrollSensitivity);
   }
 }